feat: add builder and relation generics
