+++ src/Illuminate/Database/Eloquent/HigherOrderBuilderProxy.php
@@ -10,7 +10,7 @@ class HigherOrderBuilderProxy
     /**
      * The collection being operated on.
      *
-     * @var \Illuminate\Database\Eloquent\Builder
+     * @var \Illuminate\Database\Eloquent\Builder<*>
      */
     protected $builder;
 
@@ -24,7 +24,7 @@ class HigherOrderBuilderProxy
     /**
      * Create a new proxy instance.
      *
-     * @param  \Illuminate\Database\Eloquent\Builder  $builder
+     * @param  \Illuminate\Database\Eloquent\Builder<*>  $builder
      * @param  string  $method
      * @return void
      */
