+++ src/Illuminate/Database/Eloquent/MassPrunable.php
@@ -39,7 +39,7 @@ trait MassPrunable
     /**
      * Get the prunable model query.
      *
-     * @return \Illuminate\Database\Eloquent\Builder
+     * @return \Illuminate\Database\Eloquent\Builder<static>
      */
     public function prunable()
     {
