+++ src/Illuminate/Database/Eloquent/SoftDeletingScope.php
@@ -14,8 +14,10 @@ class SoftDeletingScope implements Scope
     /**
      * Apply the scope to a given Eloquent query builder.
      *
-     * @param  \Illuminate\Database\Eloquent\Builder  $builder
-     * @param  \Illuminate\Database\Eloquent\Model  $model
+     * @template TModel of \Illuminate\Database\Eloquent\Model
+     *
+     * @param  \Illuminate\Database\Eloquent\Builder<TModel>  $builder
+     * @param  TModel  $model
      * @return void
      */
     public function apply(Builder $builder, Model $model)
@@ -26,7 +28,7 @@ class SoftDeletingScope implements Scope
     /**
      * Extend the query builder with the needed functions.
      *
-     * @param  \Illuminate\Database\Eloquent\Builder  $builder
+     * @param  \Illuminate\Database\Eloquent\Builder<*>  $builder
      * @return void
      */
     public function extend(Builder $builder)
@@ -47,7 +49,7 @@ class SoftDeletingScope implements Scope
     /**
      * Get the "deleted at" column for the builder.
      *
-     * @param  \Illuminate\Database\Eloquent\Builder  $builder
+     * @param  \Illuminate\Database\Eloquent\Builder<*>  $builder
      * @return string
      */
     protected function getDeletedAtColumn(Builder $builder)
@@ -62,7 +64,7 @@ class SoftDeletingScope implements Scope
     /**
      * Add the restore extension to the builder.
      *
-     * @param  \Illuminate\Database\Eloquent\Builder  $builder
+     * @param  \Illuminate\Database\Eloquent\Builder<*>  $builder
      * @return void
      */
     protected function addRestore(Builder $builder)
@@ -77,7 +79,7 @@ class SoftDeletingScope implements Scope
     /**
      * Add the restore-or-create extension to the builder.
      *
-     * @param  \Illuminate\Database\Eloquent\Builder  $builder
+     * @param  \Illuminate\Database\Eloquent\Builder<*>  $builder
      * @return void
      */
     protected function addRestoreOrCreate(Builder $builder)
@@ -94,7 +96,7 @@ class SoftDeletingScope implements Scope
     /**
      * Add the create-or-restore extension to the builder.
      *
-     * @param  \Illuminate\Database\Eloquent\Builder  $builder
+     * @param  \Illuminate\Database\Eloquent\Builder<*>  $builder
      * @return void
      */
     protected function addCreateOrRestore(Builder $builder)
@@ -111,7 +113,7 @@ class SoftDeletingScope implements Scope
     /**
      * Add the with-trashed extension to the builder.
      *
-     * @param  \Illuminate\Database\Eloquent\Builder  $builder
+     * @param  \Illuminate\Database\Eloquent\Builder<*>  $builder
      * @return void
      */
     protected function addWithTrashed(Builder $builder)
@@ -128,7 +130,7 @@ class SoftDeletingScope implements Scope
     /**
      * Add the without-trashed extension to the builder.
      *
-     * @param  \Illuminate\Database\Eloquent\Builder  $builder
+     * @param  \Illuminate\Database\Eloquent\Builder<*>  $builder
      * @return void
      */
     protected function addWithoutTrashed(Builder $builder)
@@ -147,7 +149,7 @@ class SoftDeletingScope implements Scope
     /**
      * Add the only-trashed extension to the builder.
      *
-     * @param  \Illuminate\Database\Eloquent\Builder  $builder
+     * @param  \Illuminate\Database\Eloquent\Builder<*>  $builder
      * @return void
      */
     protected function addOnlyTrashed(Builder $builder)
