+++ types/Database/Eloquent/Model.php
@@ -2,5 +2,6 @@
 
 use function PHPStan\Testing\assertType;
 
-$factory = User::factory();
-assertType('Illuminate\Database\Eloquent\Factories\Factory<User>', $factory);
+assertType('Illuminate\Database\Eloquent\Factories\Factory<User>', User::factory());
+
+assertType('Illuminate\Database\Eloquent\Builder<User>', User::query());
